+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 1. [#275](https://github.com/influxdata/influxdb-client-js/pull/275): Add tree shaking optimizations.
 1. [#276](https://github.com/influxdata/influxdb-client-js/pull/276): Allow to notify about write success.
 1. [#280](https://github.com/influxdata/influxdb-client-js/pull/280): Upgrade to the latest node v14 LTS.
+1. [#283](https://github.com/influxdata/influxdb-client-js/pull/283): Allow to specify custom HTTP headers in WriteApi or QueryApi.
 
 ## 1.8.0 [2020-10-30]